rename libdiskencryption to libsecurity
