+++ CMakeLists.txt
@@ -72,7 +72,7 @@ add_subdirectory(libsync)
 add_subdirectory(libledger)
 add_subdirectory(librpc)
 add_subdirectory(libinitializer)
-add_subdirectory(libdiskencryption)
+add_subdirectory(libsecurity)
 
 
 if (TESTS)
